requirements.txt
psycopg2-binary
pandas
sqlalchemy
apache-airflow
plotly-express